Cleaning up migrator and jobsdb_migrations into separate export and import files for each. Added checkpoints for setups
